General Project Cleanup
Reorganization, comments, and refactoring, oh my!
